UPDATE: Cascadia_Inverter_can
VCU-1

Verified SDP code for testing Cascadia Inverter, on Feb 17th.
Perfectly runs the propulsion system with the command message and reads the 7 broadcasted. messages.

TO-DO: 
1. Fix can node allocation for each inverter. *Urgent*
In this commit, the inverter for the left side of the vehicle is declared the right side upon initializing CAN communication.
This issue shall be resolved when testing the both inverters at once.

2. Implement reading additional broadcasted messages if needed.